계산하기위한 동작으로 ScoreCalculator객체로 분리
